The Homepage
You can edit the info per planet and can change the hyperlink to the webpage associated to that planet which goes to a separate tab. 
The external CSS sheets the pages for the nighttime and the day time, all we gotta do is change the link from 'nigh' to 'day' in the header.